Implemented new form of agent_post() and agent_poll(). Now passes 1 byte type, 2 byte length of header, header, and data which can now be binary. Types greater than 127 should now be considered user defined.

+++ libraries/support/convertdef.h
@@ -469,7 +469,7 @@ struct tlestruc
 typedef struct
 {
 	//! Number of positions
-	int32_t count;
+	size_t count;
 	//! Time step in Modified Julian Days
 	double dt;
 	//! Array of positions
